Fix population service
TRequirementRef.getRef() does not return a QName to a req/cap but returns a TNodeTemplate. That means that you cannot just add reqs/caps to the boundary definitions, you have to reference them. This reference is just a reference to a node - the boundary definitions will mirror all reqs/caps from the referenced node.