Reimplement sorting service
Use Kahn's algorithm.
Extend it to also support unfinished- and cyclic graphs. For example having unfulfilled requirements or cyclic dependencies would normally output an error. In such a case the algorithm now chooses a random node to continue with, ignores the requirements of that node, reruns Kahn's algorithm, and repeats until no nodes are left.